Не доделал
лестницы в отдельную библиотеку, массив лестниц, чтение массива из файла(тут проблема)
