+++ Lib/Map.txt
@@ -12,6 +12,9 @@ p.x = 300
 p.y = 100
 p.x1 = 900
 p.y1 = 100
+stairs
+300
+300
 bomzh
 500
 200
